requirements.txt
# Dependencias principales de Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-Caching==2.1.0
Flask-Testing==0.8.1
Flask-Babel==4.0.0

# Dependencias del sistema
Werkzeug==2.3.7
SQLAlchemy==2.0.36
python-dotenv==1.0.0
Jinja2==3.1.2
WTForms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2

# Dependencias para procesamiento de imágenes y códigos QR
Pillow==10.4.0
qrcode==7.4.2

# Dependencias para servicios externos
mercadopago==2.2.0
twilio==8.5.0
openai==1.35.0
anthropic==0.64.0

# Dependencias para WebSockets y eventos
eventlet==0.33.3

# Dependencias para mejoras críticas
PyJWT==2.8.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
flask-limiter==3.5.0
sentry-sdk[flask]==1.40.0

# Dependencias para 2FA
pyotp==2.9.0

# Dependencias para optimización de assets
cssmin==0.2.0
jsmin==3.0.1

# Dependencias para monitoreo y programación
psutil==5.9.6
APScheduler==3.10.4

# Dependencias para internacionalización
Babel==2.13.1

# Dependencias para testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-json-report==1.5.0
factory-boy==3.3.0
faker==19.12.0
coverage==7.3.2

# Dependencias para backup
schedule==1.2.0

# Dependencias para testing de interfaz
selenium==4.15.2
webdriver-manager==4.0.1

# Dependencias para Fases 1-6 (Nuevas)
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
googlemaps==4.10.0
docker==6.1.3
boto3==1.28.44
sendgrid==6.10.0
stripe==6.6.0
paypalrestsdk==1.13.1
openweathermap==0.1.0
geopy==2.3.0
